refactor(naming): add missing schema and util suffixes
