+++ app/services/auth_service.py
@@ -9,8 +9,8 @@ from app.schemas.auth_schemas import (
 from app.schemas.user_schemas import UserCreateRequest
 from app.services.password_service import PasswordService
 from app.services.email_service import EmailService
-from app.utils.error_codes import ErrorCodes
-from app.utils.exceptions import AppException
+from app.utils.error_codes_utils import ErrorCodes
+from app.utils.exceptions_utils import AppException
 
 
 class AuthService:
