+++ tests/units/controllers/test_user_controller.py
@@ -10,8 +10,8 @@ from app.schemas.user_schemas import (
     UserProfileResponse,
 )
 from app.dependencies.auth_dependency import auth_dependency
-from app.utils.exceptions import AppException
-from app.utils.error_codes import ErrorCodes
+from app.utils.exceptions_utils import AppException
+from app.utils.error_codes_utils import ErrorCodes
 
 
 @pytest.fixture
